Fixing worker dev script to not overwrite env file everytime.

+++ packages/worker/scripts/dev/manage.js
@@ -4,26 +4,28 @@ const fs = require("fs")
 
 async function init() {
   const envFilePath = path.join(process.cwd(), ".env")
-  const envFileJson = {
-    SELF_HOSTED: 1,
-    PORT: 4002,
-    CLUSTER_PORT: 10000,
-    JWT_SECRET: "testsecret",
-    INTERNAL_API_KEY: "budibase",
-    MINIO_ACCESS_KEY: "budibase",
-    MINIO_SECRET_KEY: "budibase",
-    REDIS_URL: "localhost:6379",
-    REDIS_PASSWORD: "budibase",
-    MINIO_URL: "http://localhost:10000/",
-    COUCH_DB_URL: "http://budibase:budibase@localhost:10000/db/",
-    // empty string is false
-    MULTI_TENANCY: "",
+  if (!fs.existsSync(envFilePath)) {
+    const envFileJson = {
+      SELF_HOSTED: 1,
+      PORT: 4002,
+      CLUSTER_PORT: 10000,
+      JWT_SECRET: "testsecret",
+      INTERNAL_API_KEY: "budibase",
+      MINIO_ACCESS_KEY: "budibase",
+      MINIO_SECRET_KEY: "budibase",
+      REDIS_URL: "localhost:6379",
+      REDIS_PASSWORD: "budibase",
+      MINIO_URL: "http://localhost:10000/",
+      COUCH_DB_URL: "http://budibase:budibase@localhost:10000/db/",
+      // empty string is false
+      MULTI_TENANCY: "",
+    }
+    let envFile = ""
+    Object.keys(envFileJson).forEach(key => {
+      envFile += `${key}=${envFileJson[key]}\n`
+    })
+    fs.writeFileSync(envFilePath, envFile)
   }
-  let envFile = ""
-  Object.keys(envFileJson).forEach(key => {
-    envFile += `${key}=${envFileJson[key]}\n`
-  })
-  fs.writeFileSync(envFilePath, envFile)
 }
 
 // if more than init required use this to determine the command type
